proposal: use associated types for better type-safety with opaque pointers and targets
By declaring which types are the opaque pointer and the pointee, we can ensure `from_opaque` creates the correct types, instead of needing to be chosen by the user. It's also a bit more ergonomic as we don't need to turbofish the single opaque pointee type when deconstructing the opaque pointer.